add rspec as a development dependency. Add basic spec_helper
Yes! You should depend on rake. Because the project has a Gemfile, a specific version of rake will be included in the bundle. By using `bundle exec rake` (or if the user uses rubygems-bundler, or …, or …) a guaranteed version of the gem will be used, and this will improve stability in the long run.