rewrite `self.load_fxml_resource` to be more Ruby and less brittle
? - should this be a private method?

- Don't call Java's URL.new twice (use Ruby Classes instead)
- Don't use optional ()
- Don't call `return` explicitly
- Whitespace for readability
- Using caller[] is clever, but *HIGHLY* CONFUSING AND BRITTLE
- Remove the option to pass a numeric arg to caller[] always use 1 (0 - JRubyFX::Controller#new, 1 - subclass)
- settings[:relative_to] expect a file or nil (views are relative to file's basedir)

About caller[]
In the future, I'd like to see a MVC-approach to loading Controllers and Views, but for today caller[] works by locating the controller class definition, and searching  for a view relative to that controller. Using caller[] is a hack, and doesn't scale, but it gets the job done.